requirements.txt
python-dotenv
transformers
accelerate
scipy
pika
numpy